client/backend/requirements.txt
flask
unidecode
stanza==1.8.1
python-dotenv==1.0.1
beautifulsoup4
huggingface-hub==0.14.1
pymongo==4.6.2
sentence-transformers==2.2.2
transformers==4.28.1
html5lib==1.1